+++ tests/testthat/test-range-dataset.R
@@ -1,7 +1,5 @@
 context("range dataset")
 
-source("utils.R")
-
 test_succeeds("range_dataset creates a dataset", {
 
   dataset <- range_dataset(from = 1, to = 11) %>% dataset_batch(10)
@@ -20,10 +18,11 @@ test_succeeds("range_dataset creates a dataset", {
 })
 
 
-
-
-
-
-
-
-
+test_succeeds("random_integer_dataset creates a dataset", {
+    ds1 <- random_integer_dataset(seed=4L) %>% dataset_take(10)
+    ds2 <- random_integer_dataset(seed=4L) %>% dataset_take(10)
+    # TODO: reticulate::iterate simplify doesn't work on objects with class(x)=="numeric"
+    r1 <- reticulate::iterate(ds1, as.numeric) %>% unlist()
+    r2 <- reticulate::iterate(ds2, as.numeric) %>% unlist()
+    expect_equal(r1, r2)
+})